refactor: Add background-communication module for checking AdBlocker status and improve event listener setup in dom-manipulation
